refactor: 기물 출력 포멧팅 책임 분리
