+++ src/main/java/view/DisplayBoard.java
@@ -3,8 +3,6 @@ package view;
 import board.Board;
 import java.util.Map;
 import pieces.Piece;
-import pieces.PieceType;
-import pieces.Side;
 import position.Column;
 import position.Position;
 import position.Row;
@@ -14,20 +12,6 @@ public class DisplayBoard {
     private static final int MAX_ROW = 9;
     private static final int MAX_COLUMN = 8;
     private static final String EMPTY_SYMBOL = "..";
-
-    private static final Map<PieceType, String> PIECE_SYMBOLS = Map.of(
-        PieceType.CHA, "CH",
-        PieceType.MA, "MA",
-        PieceType.SANG, "SG",
-        PieceType.SA, "SA",
-        PieceType.GUNG, "GU",
-        PieceType.PO, "PO",
-        PieceType.JOL_BYEONG, "JB"
-    );
-
-    private static final String ANSI_RESET = "\u001B[0m";
-    private static final String ANSI_RED = "\u001B[31m";
-    private static final String ANSI_GREEN = "\u001B[32m";
 
     private DisplayBoard() {
     }
@@ -65,26 +49,7 @@ public class DisplayBoard {
         if (piece == null) {
             return " " + EMPTY_SYMBOL + " ";
         }
-        String symbol = symbolOf(piece);
-        return " " + colorize(piece, symbol) + " ";
-    }
-
-    private static String colorize(Piece piece, String symbol) {
-        if (piece.isSameSide(Side.HAN)) {
-            return ANSI_RED + symbol + ANSI_RESET;
-        }
-        if (piece.isSameSide(Side.CHO)) {
-            return ANSI_GREEN + symbol + ANSI_RESET;
-        }
-        return symbol;
-    }
-
-    private static String symbolOf(Piece piece) {
-        String symbol = PIECE_SYMBOLS.get(piece.getType());
-        if (symbol == null) {
-            throw new IllegalArgumentException("출력할 수 없는 기물 타입입니다. type=" + piece.getType());
-        }
-        return symbol;
+        return " " + DisplayPiece.symbolOf(piece) + " ";
     }
 
     private static String padLeft(String value, int size) {
